Refactor first lol command to other file
