+++ cogs/apikeys.py
@@ -1,0 +1,3 @@
+token = 'MTEwMzM1ODMwNzU4MzkzNDQ3NA.GrhmtH.hGWXIqpRk2dbnBaBYfWCwLEFep5vVZYKcaUNk4'
+
+league_of_legendsAPI = 'RGAPI-3d31b2de-8770-4d72-88b4-d674fb12bf3b'
